[query] move `configureLogging` out of `HailContext`
